Implementation boilerplate.
Some things to note:

The current node structure is written in such a way that it is traversed in the reverse way to the way the Context interface describes it - that is, instead of starting at the bottom and traversing up, it starts at the top and traverses down, ending with the GlobalComponentTransitNode.

Things that will be added:

LocalComponentTransitNode needs support for custom wiring. This involves adding other contexts to its search path, because there should only be one global transit node. Further, it is in LocalComponentTransitNode where restrictions such as classloader local-ness should be implemented.

I'll put more work into refining the structure and design within the next few weeks, possibly making the node format conform to the Context documentation.